+++ sdk/appconfiguration/azure-appconfiguration/tests/test_azure_appconfiguration_client_aad.py
@@ -837,3 +837,67 @@ class TestAppConfigurationClientAAD(AppConfigTestCase):  # pylint: disable=too-m
         assert new_sent.filters[2]["parameters"]["Audience"]["DefaultRolloutPercentage"] == 100
 
         client.delete_configuration_setting(new_sent.key)
+
+    @AppConfigPreparer()
+    @recorded_by_proxy
+    def test_check_configuration_settings_by_page_etag(self, appconfiguration_endpoint_string):
+        # response header <x-ms-content-sha256> and <x-ms-date> are missing in python38.
+        set_custom_default_matcher(compare_bodies=False, excluded_headers="x-ms-content-sha256,x-ms-date")
+        self.set_up(appconfiguration_endpoint_string, is_aad=True)
+        # prepare 200 configuration settings
+        for i in range(200):
+            self.client.set_configuration_setting(
+                ConfigurationSetting(
+                    key=f"sample_key_{str(i)}",
+                    label=f"sample_label_{str(i)}",
+                )
+            )
+        # there will have 2 pages while listing, there are 100 configuration settings per page.
+
+        # get page etags using check (HEAD request)
+        match_conditions = []
+        items = self.client.check_configuration_settings(key_filter="sample_key_*", label_filter="sample_label_*")
+        iterator = items.by_page()
+        for _ in iterator:
+            etag = iterator.etag
+            match_conditions.append(etag)
+
+        # monitor page updates without changes - only changed pages will be yielded
+        items = self.client.check_configuration_settings(key_filter="sample_key_*", label_filter="sample_label_*")
+        iterator = items.by_page(match_conditions=match_conditions)
+        changed_pages = list(iterator)
+
+        # No pages should be yielded since nothing changed
+        assert len(changed_pages) == 0
+
+        # do some changes
+        self.client.set_configuration_setting(
+            ConfigurationSetting(
+                key="sample_key_201",
+                label="sample_label_202",
+            )
+        )
+        # now we have three pages, 100 settings in first two pages and 1 setting in the last page
+
+        # get page etags after updates using check (HEAD request)
+        new_match_conditions = []
+        items = self.client.check_configuration_settings(key_filter="sample_key_*", label_filter="sample_label_*")
+        iterator = items.by_page()
+        for _ in iterator:
+            etag = iterator.etag
+            new_match_conditions.append(etag)
+
+        assert match_conditions[0] == new_match_conditions[0]
+        assert match_conditions[1] != new_match_conditions[1]
+        assert match_conditions[2] != new_match_conditions[2]
+        assert len(new_match_conditions) == 3
+
+        # monitor pages after updates - only changed pages will be yielded
+        items = self.client.check_configuration_settings(key_filter="sample_key_*", label_filter="sample_label_*")
+        iterator = items.by_page(match_conditions=new_match_conditions)
+        changed_pages = list(iterator)
+        # Should yield 0 pages
+        assert len(changed_pages) == 0
+
+        # clean up
+        self.tear_down()